CLOiSim-2.1.0
[Major Changes]

* Upgrade Unity editor version
-> 2020.2.3f1

* Enable parsing SDF Actor

* import and implement actors
  - set skin
  - set animation
  - set script

* modify SDF2Unity Mesh script using assimp-net

* Handling Bone and tangent for mesh when import meshes

* Fix bone list and bind pose matrix in SkinnedMeshRenderer for actor
- bone index for bind pose is important
- set the bind poses and bon weights after combine meshes

* Separate mesh handling methods in SDF2Unity into MehLoader

* Code refactoring for Assimp Mesh Loader 
  - fix roation for skin

* Parse and implement Animation in actor sdf element

* Import animation clip for animation component

* code modified for animation curver in animation clip

* Fix rotation for animation bone axis

* fix and code refactoring for loading animation method

* apply loop pose for legacy animaion clip
  - it does not add x curves if node is root

* apply script for actor

* add capsule collider for actor

* Fix the type of SDF element in actor
  : bool -> double

* Sorting waypoint list by time when parsing actor element

* Apply following waypoint for actor

* Modify Transform Gizmo for actor
  - only support tranlate the actor with trajectories in script
  - suppport translate and rotation for the actor with trajectories


[Minor Changes]

* Print error message on UI when it failed to load document(world file)

* Update app version info in Project settings
  : 2.0.1 -> 2.1.0

* Update Packages version
  - searcher: 4.3.1 -> 4.3.2

* change type of Time to double
  -> Unity support  .....timeAsDouble

* Modify textfield property in object spawning
  - text alignment

* Code refactoring for MeshLoader
  - separate bone loading

* Change file name
  : ProdeceduralMesh -> ProceduralMesh (typo)

* Move Assimp.XXXXX.cs files into Tools/Mesh/*

* Change order of calling method in ProceduralMesh
  - optimize mesh before recalculate others...


[Bug fix]

* Fix properties of simulation display
  - bug fix height position for textbox of event and simulation message when resizing windows
  - resize width of event message

* bug fix in SDF parser
  - GetValues() method had wrong xpath

* Change order of calling method in SDF.Implement.Collision/Visual
  - optimize mesh before recalculate others...

+++ Assets/Scripts/Devices/DeviceHelper.cs
@@ -21,7 +21,7 @@ public partial class DeviceHelper
 
 			if (clock == null)
 			{
-					clock = coreObject.AddComponent<Clock>();
+				clock = coreObject.AddComponent<Clock>();
 			}
 		}
 
@@ -56,6 +56,7 @@ public partial class DeviceHelper
 		}
 	}
 
+
 	public static string GetPartName(in GameObject targetObject)
 	{
 		return GetModelName(targetObject, true);
@@ -73,7 +74,7 @@ public partial class DeviceHelper
 
 		var timeNow = (useRealTime) ? GetGlobalClock().GetRealTime() : GetGlobalClock().GetSimTime();
 		msgTime.Sec = (int)timeNow;
-		msgTime.Nsec = (int)((timeNow - (float)msgTime.Sec) * (float)1e+9);
+		msgTime.Nsec = (int)((timeNow - (double)msgTime.Sec) * (double)1e+9);
 	}
 
 	public static void SetVector3d(messages.Vector3d vector3d, in Vector3 position)
